Refactor auto/options module
